requirements.txt
requests
python-telegram-bot==20.1
python-dotenv
httpx
datetime
pytz
beautifulsoup4